+++ xbmc/cores/VideoPlayer/DVDDemuxers/DVDDemuxBXA.cpp
@@ -164,7 +164,7 @@ std::vector<CDemuxStream*> CDVDDemuxBXA::GetStreams() const
   return streams;
 }
 
-int CDVDDemuxBXA::GetNrOfStreams() const
+size_t CDVDDemuxBXA::GetNrOfStreams() const
 {
   return (m_stream == NULL ? 0 : 1);
 }
