+++ xbmc/cores/VideoPlayer/PTSTracker.h
@@ -23,7 +23,7 @@ class CPtsTracker
     void   Flush(); //flush the saved pattern and the ringbuffer
     void   ResetVFRDetection(void);
 
-    int    GetPatternLength() { return m_patternlength;            }
+    size_t GetPatternLength() { return m_patternlength; }
     double GetFrameDuration() { return m_frameduration;            }
     double GetMaxFrameDuration(void) { return m_maxframeduration;  }
     double GetMinFrameDuration(void) { return m_minframeduration;  }
@@ -34,13 +34,15 @@ class CPtsTracker
     double m_prevpts;                //last pts added
     double m_diffring[DIFFRINGSIZE]; //ringbuffer of differences between pts'
     int    m_ringpos;                //position of last diff added to ringbuffer
-    int    m_ringfill;               //how many diffs we have in the ringbuffer
-    double GetDiff(int diffnr);      //gets diffs from now to the past
+    size_t m_ringfill; //how many diffs we have in the ringbuffer
+    double GetDiff(size_t diffnr); //gets diffs from now to the past
 
     void GetPattern(std::vector<double>& pattern);     //gets the current pattern
 
     static bool MatchDiff(double diff1, double diff2); //checks if two diffs match by MAXERR
-    static bool MatchDifftype(int diffs1[], int diffs2[], int nrdiffs); //checks if the difftypes match
+    static bool MatchDifftype(int diffs1[],
+                              int diffs2[],
+                              size_t nrdiffs); //checks if the difftypes match
 
     //checks if the current pattern matches with the saved m_pattern with offset m_patternpos
     bool CheckPattern(std::vector<double>& pattern);
@@ -52,7 +54,7 @@ class CPtsTracker
     double m_maxframeduration;     //Max value detected for frame duration (for VFR files case)
     double m_minframeduration;     //Min value detected for frame duration (for VFR files case)
     bool m_haspattern;             //for the log and detecting VFR files case
-    int m_patternlength;           //for the codec info
+    size_t m_patternlength; //for the codec info
     int m_VFRCounter;              //retry counter for VFR detection
     int m_patternCounter;
     std::string GetPatternStr();   //also for the log
